[SQL] Optimize temporal filters with disjunctions
